Move ALLOW_FIXTURES from mypy.test.config to mypy.defaults to please python 3.5
